edmc prompt cleanup, trade alert message feedback

+++ MissionAlertBot.py
@@ -849,14 +849,15 @@ TEXT GEN FUNCTIONS
 
 def txt_create_discord(carrier_data, mission_type, commodity, station, system, profit, pads, demand, eta_text, mission_temp_channel_id, edmc_off):
     discord_channel = f"<#{mission_temp_channel_id}>" if mission_temp_channel_id else f"#{carrier_data.discord_channel}"
-    discord_text = f"{discord_channel} {'load' if mission_type == 'load' else 'unload'}ing " \
-                   f"{commodity.name} " \
-                   f"{'from' if mission_type == 'load' else 'to'} **{station.upper()}** station in system " \
-                   f"**{system.upper()}** : {profit}k per unit profit : "\
-                   f"{demand} {'demand' if mission_type == 'load' else 'supply'} : {pads.upper()}-pads" \
-                   f".{eta_text}"
-    if edmc_off:
-        discord_text += " **EDMC OFF**"
+    discord_text = (
+        f"{'**★ EDMC-OFF MISSION! ★** : ' if edmc_off else ''}"
+        f"{discord_channel} {'load' if mission_type == 'load' else 'unload'}ing "
+        f"{commodity.name} "
+        f"{'from' if mission_type == 'load' else 'to'} **{station.upper()}** station in system "
+        f"**{system.upper()}** : {profit}k per unit profit : "
+        f"{demand} {'demand' if mission_type == 'load' else 'supply'} : {pads.upper()}-pads"
+        f".{eta_text}"
+    )
     return discord_text
 
 
@@ -940,7 +941,7 @@ async def on_guild_channel_pins_update(channel, last_pin):
     """
     messages = await channel.history(limit=200).flatten()
     for msg in messages:
-        if msg.type is discord.MessageType.pins_add:
+        if msg.type is discord.MessageType.pins_add and msg.author == bot.user:
             await msg.delete()
 
 
@@ -1233,21 +1234,8 @@ async def gen_mission(ctx, carrier_name_search_term: str, commodity_search_term:
         await message_gen.delete()
         print("Output check displayed")
 
-        embed = discord.Embed(title="Is this an EDMC OFF mission?",
-                            description="Should haulers be instructed to turn EDMC and other 3rd party tools off?\ny/n",
-                            color=constants.EMBED_COLOUR_QU)
-        edmc_confirm = await ctx.send(embed=embed)
-        try:
-            check_characters = 'ny'
-            edmc_msg = await bot.wait_for("message", check=check_confirm, timeout=30)
-            if 'y' == edmc_msg.content.lower().strip():
-                edmc_off = True
-        except asyncio.TimeoutError:
-            await default_timeout_message('edmc_prompt')
-            return
-
         embed = discord.Embed(title="Where would you like to send the alert?",
-                            description="(**d**)iscord, (**r**)eddit, (**t**)ext for copy/pasting or (**x**) to cancel\n"
+                            description="(**d**)iscord, (**e**)dmc OFF, (**r**)eddit, (**t**)ext for copy/pasting or (**x**) to cancel\n"
                             "Use (**n**) to also notify PTN Haulers.",
                             color=constants.EMBED_COLOUR_QU)
         embed.set_footer(text="Enter all that apply, e.g. **drn** will send alerts to Discord and Reddit and notify PTN Haulers.")
@@ -1255,16 +1243,15 @@ async def gen_mission(ctx, carrier_name_search_term: str, commodity_search_term:
         print("Prompted user for alert destination")
 
         try:
-            check_characters = 'drtnx'
+            check_characters = 'dertnx'
             msg = await bot.wait_for("message", check=check_confirm, timeout=30)
+            edmc_off = True if "e" in msg.content.lower() else False
 
             if "x" in msg.content.lower():
                 # immediately stop if there's an x anywhere in the message, even if there are other proper inputs
                 message_cancelled = await ctx.send("**Mission creation cancelled.**")
                 await msg.delete()
                 await message_confirm.delete()
-                await edmc_confirm.delete()
-                await edmc_msg.delete()
                 print("User cancelled mission generation")
                 return
 
@@ -1414,7 +1401,7 @@ async def gen_mission(ctx, carrier_name_search_term: str, commodity_search_term:
                             color=constants.EMBED_COLOUR_DISCORD)
                 await ctx.send(embed=embed)
 
-            if edmc_off and "d" in msg.content.lower():
+            if "e" in msg.content.lower() and "d" in msg.content.lower():
                 print('Sending EDMC OFF messages to haulers')
                 embed = discord.Embed(title='PLEASE STOP ALL 3RD PARTY SOFTWARE: EDMC, EDDISCOVERY, ETC',
                         description=("Maximising our haulers' profits for this mission means keeping market data at this station"
@@ -1429,12 +1416,11 @@ async def gen_mission(ctx, carrier_name_search_term: str, commodity_search_term:
 
                 embed = discord.Embed(title=f"EDMC OFF messages sent", description='Reddit posts will be skipped',
                             color=constants.EMBED_COLOUR_DISCORD)
+                await ctx.send(embed=embed)
 
                 print('Reacting to #official-trade-alerts message with EDMC OFF')
                 for r in ["🇪","🇩","🇲","🇨","📴"]:
                     await trade_alert_msg.add_reaction(r)
-
-                await ctx.send(embed=embed)
 
         except asyncio.TimeoutError:
             await default_timeout_message('notification_type_decision')
@@ -1453,8 +1439,6 @@ async def gen_mission(ctx, carrier_name_search_term: str, commodity_search_term:
         try:
             await msg.delete()
             await message_confirm.delete()
-            await edmc_confirm.delete()
-            await edmc_msg.delete()
         except Exception as e:
             print(f"Error during clearup: {e}")
             await ctx.send(f"Oops, error detected: {e}\nAttempting to continue with mission gen.")
